Move API invocation code in VMware drivers to OSLO
There is lot of common code between VMware drivers for nova and cinder;
for e.g., session management and VIM API invocation code. The new
VMware driver for glance and the telemetry agent for vSphere are also
planning to use this common code. This change moves the classes in
VMware drivers for session and API call management calls into OSLO.

The api module in this change contains classes to invoke VIM APIs.
It supports automatic session re-establishment and retry of API
invocations in case of connection problems or server API call overload.

Change-Id: I74e2be632631f83a97e26c03e33f1d9ae7632e27
Implements: blueprint vmware-api