Remove __del__ usage in oslo.vmware driver
The driver uses __del__ in which may cause spurious errors.
python docs detail the caveats of using __del__.

In order to enable the calling application to be able to close
existing sessions, a new method has been added: logout.
This will be used to terminate and logout of the existing session.

Change-Id: I94482b79cc451d027cb853a6d0775e4da189cb7e