Switch off caching to prevent cache poisoning by local attacker
suds package uses a well-known /tmp/suds directory by default to
store pickled objects. This could be subject to attacks. suds project
is un-maintained and some distros have a local patch that uses
mkdtemp() but still leak directories. It's best to remove the cache
completely. Note that if we really need, we could add a memory
based cached later. we should deal with this security issue first.

Closes-Bug: #1341954

Change-Id: Ieec9d99aa674adf5cbc9be924fef3856cf4e5d66