Add API to enable calling module to register an exception
In certain cases the application, for example Nova, may need to treat
a VC exception that is currently not supported by oslo.vmware.

This will enable the application to dynamically add exceptions and then
remove them when oslo.vmware is updated to support the new exception.

The patch also fixes a typo with the exception fault handling.

Change-Id: I332a4ff2ac2c6d4e6ac88bbd067ef0d4e659ffa3