Fix handling of fault details
Fix regression caused by adding a new transport based on 'requests'
in commit 4bd0b4cf. When a Fault message is returned from the backed,
the 'Envelope' and 'Body' tags are stripped and the root element becomes
'Fault'. The path to the 'detail' string becomes '/detail' instead of
'/Envelope/Body/Fault/detail'. The stripping is done in:

suds/bindings/binding.py#detect_fault()

which was not executed when the old http transport was used.

Change-Id: I1b99d37fdbcbf5ea05e4823e408468a06e230ccf