Supress error logs when exception is thrown
There are many cases when exceptions from the backend are expected and
these exceptions must not be logged as errors. For example it is a
common pattern to catch DuplicateName/AlreadyExists exceptions and
ignore them in the cases when multiple threads are trying to perform the
same operation. Currently there is no way for clients to prevent logging
these exceptions as errors and the proper way to fix this is to just
raise the exception in oslo and let the client decide if it must be
logged as an error or not.

Closes-Bug: #1377927
Change-Id: I5309cad5b0eeb380387eb3051d80c59b783561e8