Do not log when reraising an exception
It should be up to the caller to use the log level error
when the exception is reraised. For example, if the call
to a VIM api fails, _poll_task is currently throwing a big
red block even if the caller ignores the exception.

This patch removes the logging from _poll_task.

Change-Id: I5bceb41c9a24dca45ff8689daee0eeafa7da839c