set default python to python3
Set the default python to python3 except for the py27 environment. We
have to set that explicitly to override the new default.

Because flake8 now runs under python 3 there is no built-in name
'file'. Instead we can use the same symbols from the io module to
build the spec for mocking out the open() call. The old version of
that code that used to only run under python 3 used a private version
of the io module, so that is also corrected.

Change-Id: I2491e65736d2c79f1134eec6371967c39873891b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>